abstracted routing functional logic out into a new class
